- rust implementation of maxflow benchmark - espprc adjustment in most implementations: rounding is done from the wrapping function, the espprc algorithm returns a floating point number

+++ implementations/c++/espprc.h
@@ -171,7 +171,7 @@ public:
 	maxlen_ = maxlen;
     }
 
-    int solve(){
+    double solve(){
 	// set class variables
 	Label<T>::nresources = nresources_;
 	Label<T>::nnodes = nnodes_;
@@ -233,7 +233,7 @@ public:
 		bestcost = labels[0][i]->cost();
 	    }
 	}
-	return (int) bestcost;
+	return bestcost;
     }
 };
